+++ R/plot.gexf.r
@@ -1,8 +1,8 @@
-#' @name gexf-methods
-#' @export
-plot.gexf <- function(x, EdgeType=NULL, output.dir=NULL,...){  
-  # WIP
-
-  .Defunct("Gephi for visualizing")  
-}
-
+#' #' @name gexf-methods
+#' #' @export
+#' plot.gexf <- function(x, EdgeType=NULL, output.dir=NULL,...){  
+#'   # WIP
+#' 
+#'   .Defunct("Gephi for visualizing")  
+#' }
+#'